+++ jitsi-meet/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Creating regular pot jail for jitsi-meet as alternative to jitsi-meet-nomad
 * This version includes consul, node_exporter, syslog-ng
 * Fix cert renew script name
+* Fix jitsi config and typo in service start
 
 ---